Add files via upload
For this project, you must write 3 programs in C. These exercises are extensions of the exercises assigned in the previous module. For the first two exercises, the extension consists of going from a unique character to a string of characters. For the third exercise, you must get “closer” to the computer and process the numbers as 0s and 1s using ONLY logic bitwise operations, rather than arithmetic operations.

For all these exercises, you cannot use built-in functions that perform these operations. If in doubt, check with your instructor.

Download the Programming Project.pdf