Add files via upload
For this assignment, you must write three programs in the high-level language you learned in CPSC 1213. The objective of this assignment is to prepare your environment to program in assembly and write your first assembly program in VisualStudio. To complete, you must have installed:

Windows (7 or later, if you do not have it)
Visual Studio IDE configured to compile assembly x86 programs
A working sample *project (provided by the author of the textbook), and your first assembly *project.
Refer to the following detailed instructions in the  Programming Assignment预览文档.
If you have questions, check with your instructor.

esComponent-callout
In order to complete Programming Assignment 2, you will need to download and review the following files (See videos at the bottom to help you if necessary):

How to Install Visual Studio预览文档 (Read and Install)
How to Create a New Project in VisualStudio预览文档 (Read)
如何入门  （链接到外部网站。）链接到外部网站。(Read Website)
Irvine.zip (Install and Run)
Project32.zip (Install and Run)
TemplateProject.zip (Follow instructions on the Programming Assignment 2 document)